Increase backoff limit for failing graph syncs
When running graph-sync-jobs in the cluster, there was an issue when
graph-syncs did not sync the document and retried just once. If the second
retry failed, they went to failed state and were deleted by OpenShift. Let's
increase backoff limit to retry syncs multiple times to get into a successful
state or to have reported issues.